Updated GSL sources to 2.8. Note that MSVC is not working yet.

+++ specfunc/gamma.c
@@ -1611,7 +1611,7 @@ int gsl_sf_choose_e(unsigned int n, unsigned int m, gsl_sf_result * result)
           prod *= tk;
         }
         result->val = prod;
-        result->err = 2.0 * GSL_DBL_EPSILON * prod * fabs(n-m);
+        result->err = 2.0 * GSL_DBL_EPSILON * prod * fabs((double) (n-m));
         return GSL_SUCCESS;
       }
     else
